+++ core/src/components/page/Page.vue
@@ -23,7 +23,7 @@
 			<pageTitle />
 			<sketches class="collectionContainer" />
 			<textBoxes class="collectionContainer" />
-			<shapes class="collectionContainer" ref="shapes" />
+			<shapes class="collectionContainer" />
 			<Moveable
 				class="moveable"
 				v-bind="moveableOptions"
@@ -59,6 +59,10 @@ import { drawing } from "../../mixins/editingModes/drawing.js";
 import { addTextBox } from "../../mixins/editingModes/addTextBox.js";
 import { addShape } from "../../mixins/editingModes/addShape.js";
 import { domToOjectId } from "../../mixins/editingModes/domToOjectId.js";
+import { selectoEventHandlers } from "../../mixins/editingModes/selectoEventHandlers.js";
+import { moveableEventHandlers } from "../../mixins/editingModes/moveableEventHandlers.js";
+import { isMountedMixin } from "../../mixins/editingModes/isMountedMixin.js";
+
 import Sketches from "../objects/Sketches.vue";
 import textBoxes from "../objects/TextBoxes.vue";
 import shapes from "../objects/Shapes.vue";
@@ -77,29 +81,15 @@ export default {
 		VueSelecto,
 		Moveable,
 	},
-	mixins: [drawing, addTextBox, addShape, domToOjectId],
+	mixins: [drawing, addTextBox, addShape, domToOjectId, selectoEventHandlers, moveableEventHandlers, isMountedMixin],
 	data: function() {
 		return {
 			targets: [],
 			selectableTargets: [".object"],
 			body: document.body,
-			isMounted: false,
 		};
 	},
 	methods: {
-		onSelect: function(e) {
-			let selectedObjects = this.domObjectsToIds(e.selected);
-
-			this.$store.commit("updateFocusedObjects", {objects: selectedObjects,}, {module: "core" });
-
-			this.targets = e.selected;
-		},
-		onDragStart: function(e) {
-			if(!(this.editingMode === "selecting")) {
-				e.stop();
-			}
-			console.log(e);
-		},
 		pointerdown: function(event) {
 			if(this.debug) {
 				console.log("pointerdown");
@@ -204,96 +194,6 @@ export default {
 		calculatePressure: function(event) {
 			return this.selectedPencil.width * 2 * (event.pressure || 0.5);
 		},
-
-
-
-
-
-		// Moveable
-		handleDragGroupStart({ events }) {
-			for(let e of events) {
-				this.handleDragStart(e);
-			}
-		},
-		handleDragStart(e) {
-			let id = this.domShapeToId(e.target);
-			let shape = this.loadedPage.objects.shapes[id];
-
-			e.set([shape.position.x, shape.position.y]);
-		},
-		handleDragGroup({ events }) {
-			for(let e of events) {
-				console.log(e);
-				this.handleDrag(e);
-			}
-		},
-		handleDrag({ target, transform, beforeTranslate }) {
-			target.style.transform = transform;
-
-			let {id, type} = this.domObjectToId(target);
-
-			this.$store.commit("moveObject", {id, type, x: beforeTranslate[0], y: beforeTranslate[1],}, {module: "core" });
-		},
-
-		handleResizeGroupStart({ events }) {
-			for(let e of events) {
-				this.handleResizeStart(e);
-			}
-		},
-		handleResizeStart(e) {
-			let id = this.domShapeToId(e.target);
-			let shape = this.loadedPage.objects.shapes[id];
-
-			e.setOrigin(["%", "%"]);
-			e.dragStart && e.dragStart.set([shape.position.x, shape.position.y]);
-		},
-
-		handleResizeGroup({ events }) {
-			for(let e of events) {
-				this.handleResize(e);
-			}
-		},
-
-		handleResize({target, width, height, delta, drag, }) {
-			delta[0] && (target.style.width = `${width}px`);
-			delta[1] && (target.style.height = `${height}px`);
-
-			target.style.transform = drag.transform;
-
-			let {id, type} = this.domObjectToId(target);
-
-			this.$store.commit("resizeObject", {id, type, width, height,}, {module: "core" });
-			this.$store.commit("moveObject", {id, type, x: drag.beforeTranslate[0], y: drag.beforeTranslate[1],}, {module: "core" });
-		},
-
-		handleRotateGroupStart({ events }) {
-			for(let e of events) {
-				this.handleRotateStart(e);
-			}
-		},
-
-		handleRotateStart(e) {
-			console.log(e);
-			let id = this.domShapeToId(e.target);
-			let shape = this.loadedPage.objects.shapes[id];
-
-			e.set(shape.position.rotation);
-		},
-
-		handleRotateGroup({ events }) {
-			for(let e of events) {
-				this.handleRotate(e);
-			}
-		},
-
-		handleRotate({ target, rotate, drag }) {
-			target.style.transform = drag.transform;
-
-			let {id, type} = this.domObjectToId(target);
-
-			this.$store.commit("rotateObject", {id, type, rotation: rotate,}, {module: "core" });
-			this.$store.commit("moveObject", {id, type, x: drag.beforeTranslate[0], y: drag.beforeTranslate[1],}, {module: "core" });
-		},
 	},
 	computed: {
 		...mapState({
@@ -312,11 +212,6 @@ export default {
 			"selectedPencil",
 			"moveableOptions",
 		]),
-	},
-	mounted() {
-		console.log(this.$refs.shapes);
-		console.log(this.$refs.zoomedContainer);
-		this.isMounted = true;
 	},
 };
 </script>
