+++ core/src/mixins/editingModes/isMountedMixin.js
@@ -1,0 +1,10 @@
+export const isMountedMixin = {
+	mounted() {
+		this.isMounted = true;
+	},
+	data: function() {
+		return {
+			isMounted: false,
+		};
+	},
+};
